renamed 2 compoents fro mock
