--- SoYummy.SPA/src/components/UserMenu/index.js
@@ -1,1 +1,0 @@
-export * from './UserMenu';
