+++ SoYummy.SPA/src/components/UserMenuMock/index.js
@@ -1,0 +1,1 @@
+export * from "./UserMenuMock";
